+++ lib/src/rendering/rough/rough.dart
@@ -4,3 +4,4 @@ export 'freedraw_renderer.dart';
 export 'path_dash_utility.dart';
 export 'rough_adapter.dart';
 export 'rough_canvas_adapter.dart';
+export 'rough_path_cache.dart';